initial commit of changing dec to long refactor

+++ src/Angor/Client/Services/ICurrencyService.cs
@@ -1,5 +1,5 @@
 public interface ICurrencyService
 {
-    Task<IReadOnlyList<string>> GetBtcValuesInPreferredCurrency(params decimal[] btcBalances);
+    Task<IReadOnlyList<string>> GetBtcValuesInPreferredCurrency(params long[] btcBalances);
     string GetCurrencySymbol(string currencyCode);
 }
